app/requirements.txt
bumpversion==0.5.3
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
Django==2.1.3
django-adminfilters==1.1.0
django-auth-ldap==1.7.0
django-ckeditor==5.6.1
django-colorfield==0.1.15
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-js-asset==1.1.0
django-mailbox==4.7.1
djangorestframework==3.9.0
filetype==1.0.1
idna==2.7
Markdown==3.0.1
Pillow==5.3.0
pprint==0.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.5
python-ldap==3.1.0
pytz==2018.7
requests==2.20.0
responses==0.10.2
rest-framework-cache==0.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.24
unicodecsv==0.14.1
redis-py-cluster
PyYAML==3.13
pika==0.12.0
robotframework-rabbitmq